requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# E:\Coding\TokenScrape\token-scrape-hfr\tok_scrape\hfr_token_scrape_v2.py: 2,3
Flask == 1.1.1

# E:\Coding\TokenScrape\token-scrape-hfr\tok_scrape\hfr_token_scrape_v2.py: 1
Mako == 1.1.4

# E:\Coding\TokenScrape\token-scrape-hfr\tok_scrape\hfr_token_scrape_v2.py: 4
aiohttp == 3.7.4.post0

# E:\Coding\TokenScrape\token-scrape-hfr\tok_scrape\hfr_token_scrape_v2.py: 8
python_dotenv == 0.17.0

# E:\Coding\TokenScrape\token-scrape-hfr\tok_scrape\hfr_token_scrape_v2.py: 9
requests == 2.22.0

# E:\Coding\TokenScrape\token-scrape-hfr\tok_scrape\hfr_token_scrape_v2.py: 5,6,7
web3 == 5.18.0
